pyproject.toml
[project]
name = "ocr-autoresearch"
version = "0.1.0"
description = "Autonomous OCR research on a tiny self-contained benchmark"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.0.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "torch>=2.5.0",
]

